refact(api): Service interfaces moved to corresponding social/fitness packages for improved readability
