+++ backend/fit-journey-api/src/main/java/hu/hm/fitjourneyapi/services/interfaces/fitness/ExerciseService.java
@@ -1,4 +1,4 @@
-package hu.hm.fitjourneyapi.services.interfaces;
+package hu.hm.fitjourneyapi.services.interfaces.fitness;
 
 import hu.hm.fitjourneyapi.dto.fitness.excercise.AbstractExerciseDTO;
 import hu.hm.fitjourneyapi.dto.fitness.excercise.ExerciseCardioSetDTO;
