+++ backend/fit-journey-api/src/main/java/hu/hm/fitjourneyapi/services/interfaces/fitness/SetService.java
@@ -1,4 +1,4 @@
-package hu.hm.fitjourneyapi.services.interfaces;
+package hu.hm.fitjourneyapi.services.interfaces.fitness;
 
 import hu.hm.fitjourneyapi.dto.fitness.set.AbstractSetDTO;
 import hu.hm.fitjourneyapi.dto.fitness.set.CardioSetDTO;
