+++ backend/fit-journey-api/src/main/java/hu/hm/fitjourneyapi/services/interfaces/social/FriendService.java
@@ -1,4 +1,4 @@
-package hu.hm.fitjourneyapi.services.interfaces;
+package hu.hm.fitjourneyapi.services.interfaces.social;
 
 import hu.hm.fitjourneyapi.dto.social.friend.FriendDTO;
